Added Datasets & Tweet Class
The datasets folder includes both the full and sample versions of the kaggle and cloudflower data sets. I did not include the full harvard data set since its too big for GitHub. I also added a module for the tweet data class. It is currently just a data class, if we decide we need methods for it, I can change it as such. I also added the project proposal pdf and tex file incase (idk, for future reference?? meh)