Added processed sample dataset and json_to_tweets() function
I wrote a new function to process the lines from the jsonl file to a tweet list, so for testing out plotting the tweets use that function to get a usable list of tweets. The processed sample dataset can be found in the file sampleHarvardProcessedTweets, let me know if its too small.  Also note that while right now this file is still pretty messy, I will work on cleaning up, commenting, and improving the usability of it in the next few days.